requirements.txt
customtkinter
darkdetect
diffusers
filelock
huggingface-hub
idna
importlib-metadata
numpy
packaging
Pillow
pyparsing
PyYAML
regex
requests
tk
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typing_extensions
urllib3
zipp
safetensors
fsspec
certifi
charset-normalizer
nvidia-cusparse-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cudnn-cu12
triton
networkx
nvidia-nccl-cu12
nvidia-nvtx-cu12
nvidia-cusolver-cu12
nvidia-cuda-runtime-cu12
nvidia-cuda-cupti-cu12
nvidia-cufft-cu12
nvidia-cublas-cu12
nvidia-curand-cu12
nvidia-nvjitlink-cu12
jinja2
sympy
mpmath